feat(): Refactor loading and configuration so that any script can easily startup seneca
