+++ src/firebase-sn.js
@@ -37,6 +37,22 @@ export default function({collections, cfg: {FIREBASE_HOST, FIREBASE_TOKEN}}) {
     this.add({role:'Firebase',model:name,cmd:'first'}, function({by, value}, respond) {
       model.first(by, value).then(result => respond(null, result))
     })
+
+    this.add({role:'Firebase',model:name,cmd:'update'}, function({key, values}, respond) {
+      model.child(key).update(values)
+        .then(() => respond(null, key))
+    })
+
+    this.add({role:'Firebase',model:name,cmd:'push'}, function({values}, respond) {
+      const key = model.push(values).key()
+      respond(null, key)
+    })
+
+    this.add({role:'Firebase',model:name,cmd:'remove'}, function({key}, respond) {
+      if(!key) { response(null, {error: 'no key'}) }
+      model.child(key).remove()
+        .then(() => respond(null, key))
+    })
   }
 
   this.add({init:'firebase-sn'}, function(args, respond) {
