+++ src/load.js
@@ -1,0 +1,14 @@
+import senecaSn from './seneca-sn'
+import Seneca from 'seneca-await'
+import cfg from './cfg'
+
+const seneca = Seneca()
+const remote = {}
+
+async function load() {
+  seneca.use(senecaSn, {cfg, remote})
+  await seneca.ready()
+  return seneca
+}
+
+export default load
